+++ src/classes/job.ts
@@ -8,6 +8,7 @@ import {
   JobJson,
   JobJsonRaw,
   MinimalJob,
+  MoveToDelayedOpts,
   MoveToWaitingChildrenOpts,
   ParentKeys,
   ParentOpts,
@@ -97,6 +98,12 @@ export class Job<
    * Timestamp when the job was created (unless overridden with job options).
    */
   timestamp: number;
+
+  /**
+   * Number of attempts when job is moved to active.
+   * @defaultValue 0
+   */
+  attemptsStarted = 0;
 
   /**
    * Number of attempts after the job has failed.
@@ -311,7 +318,10 @@ export class Job<
     }
 
     job.failedReason = json.failedReason;
-    job.attemptsMade = parseInt(json.attemptsMade || '0');
+
+    job.attemptsStarted = parseInt(json.ats || '0');
+
+    job.attemptsMade = parseInt(json.attemptsMade || json.atm || '0');
 
     job.stacktrace = getTraces(json.stacktrace);
 
@@ -430,6 +440,7 @@ export class Job<
       parentKey: this.parentKey,
       progress: this.progress,
       attemptsMade: this.attemptsMade,
+      attemptsStarted: this.attemptsStarted,
       finishedOn: this.finishedOn,
       processedOn: this.processedOn,
       timestamp: this.timestamp,
@@ -585,7 +596,10 @@ export class Job<
     );
 
     const result = await this.scripts.moveToFinished(this.id, args);
-    this.finishedOn = args[14] as number;
+    this.finishedOn = args[
+      this.scripts.moveToFinishedKeys.length + 1
+    ] as number;
+    this.attemptsMade += 1;
 
     return result;
   }
@@ -620,7 +634,7 @@ export class Job<
     let moveToFailed = false;
     let finishedOn, delay;
     if (
-      this.attemptsMade < this.opts.attempts &&
+      this.attemptsMade + 1 < this.opts.attempts &&
       !this.discarded &&
       !(err instanceof UnrecoverableError || err.name == 'UnrecoverableError')
     ) {
@@ -629,7 +643,7 @@ export class Job<
       // Check if backoff is needed
       delay = await Backoffs.calculate(
         <BackoffOptions>this.opts.backoff,
-        this.attemptsMade,
+        this.attemptsMade + 1,
         err,
         this,
         opts.settings && opts.settings.backoffStrategy,
@@ -667,7 +681,7 @@ export class Job<
         fetchNext,
       );
       (<any>multi).moveToFinished(args);
-      finishedOn = args[14];
+      finishedOn = args[this.scripts.moveToFinishedKeys.length + 1] as number;
       command = 'failed';
     }
 
@@ -692,9 +706,14 @@ export class Job<
       this.delay = delay;
     }
 
+    this.attemptsMade += 1;
+
     if (Array.isArray(result)) {
       return raw2NextJobData(result);
     }
+
+
+
   }
 
   /**
@@ -1029,14 +1048,17 @@ export class Job<
    * @param token - token to check job is locked by current worker
    * @returns
    */
-  moveToDelayed(timestamp: number, token?: string): Promise<void> {
+  async moveToDelayed(timestamp: number, token?: string): Promise<void> {
     const delay = timestamp - Date.now();
-    return this.scripts.moveToDelayed(
+    const movedToDelayed = await this.scripts.moveToDelayed(
       this.id,
       timestamp,
       delay > 0 ? delay : 0,
       token,
+      { skipAttempt: true },
     );
+
+    return movedToDelayed;
   }
 
   /**
@@ -1046,11 +1068,17 @@ export class Job<
    * @param opts - The options bag for moving a job to waiting-children.
    * @returns true if the job was moved
    */
-  moveToWaitingChildren(
+  async moveToWaitingChildren(
     token: string,
     opts: MoveToWaitingChildrenOpts = {},
   ): Promise<boolean> {
-    return this.scripts.moveToWaitingChildren(this.id, token, opts);
+    const movedToWaitingChildren = await this.scripts.moveToWaitingChildren(
+      this.id,
+      token,
+      opts,
+    );
+
+    return movedToWaitingChildren;
   }
 
   /**
@@ -1142,10 +1170,7 @@ export class Job<
     }
 
     if (`${parseInt(this.id, 10)}` === this.id) {
-      //TODO: throw an error in next breaking change
-      console.warn(
-        'Custom Ids should not be integers: https://github.com/taskforcesh/bullmq/pull/1569',
-      );
+      throw new Error('Custom Ids cannot be integers');
     }
 
     if (this.opts.priority) {