+++ tests/fixtures/fixture_processor_unrecoverable.js
@@ -9,10 +9,10 @@ const delay = require('./delay');
 
 module.exports = function (job) {
   return delay(500).then(() => {
-    if (job.attemptsMade < 2) {
+    if (job.attemptsMade < 1) {
       throw new Error('Not yet!');
     }
-    if (job.attemptsMade < 3) {
+    if (job.attemptsMade < 2) {
       throw new UnrecoverableError('Unrecoverable');
     }
   });