+++ tests/test_job.ts
@@ -1097,12 +1097,16 @@ describe('Job', function () {
       );
       await worker.waitUntilReady();
 
-      const completing = new Promise<void>(resolve => {
+      const completing = new Promise<void>((resolve, reject) => {
         worker.on(
           'completed',
           after(4, () => {
-            expect(completed).to.be.eql(['a', 'b', 'c', 'd']);
-            resolve();
+            try {
+              expect(completed).to.be.eql(['a', 'b', 'c', 'd']);
+              resolve();
+            } catch (err) {
+              reject(err);
+            }
           }),
         );
       });