+++ tests/test_pause.ts
@@ -364,7 +364,7 @@ describe('Pause', function () {
           queueName,
           async job => {
             await delay(10);
-            if (job.attemptsMade == 1) {
+            if (job.attemptsMade == 0) {
               await queue.pause();
               throw new Error('Not yet!');
             }