+++ tests/test_worker.ts
@@ -465,6 +465,95 @@ describe('workers', function () {
     await worker.close();
   }).timeout(8000);
 
+  it('do not call moveToActive more than concurrency factor + 1', async () => {
+    const numJobs = 57;
+    const concurrency = 13;
+    let completedJobs = 0;
+    const worker = new Worker(
+      queueName,
+      async job => {
+        expect(job.data.foo).to.be.equal('bar');
+        await delay(250);
+      },
+      { connection, prefix, concurrency },
+    );
+    await worker.waitUntilReady();
+
+    // Add spy to worker.moveToActive
+    const spy = sinon.spy(worker, 'moveToActive');
+    const bclientSpy = sinon.spy(
+      await worker.blockingConnection.client,
+      'bzpopmin',
+    );
+
+    for (let i = 0; i < numJobs; i++) {
+      const job = await queue.add('test', { foo: 'bar' });
+      expect(job.id).to.be.ok;
+      expect(job.data.foo).to.be.eql('bar');
+    }
+
+    expect(bclientSpy.callCount).to.be.equal(1);
+
+    await new Promise<void>((resolve, reject) => {
+      worker.on('completed', (_job: Job, _result: any) => {
+        completedJobs++;
+        if (completedJobs == numJobs) {
+          resolve();
+        }
+      });
+    });
+
+    // Check moveToActive was called only concurrency times
+    expect(spy.callCount).to.be.equal(concurrency + 1);
+    expect(bclientSpy.callCount).to.be.equal(3);
+
+    await worker.close();
+  });
+
+  it('do not call moveToActive more than number of jobs + 1', async () => {
+    const numJobs = 50;
+    let completedJobs = 0;
+    const worker = new Worker(
+      queueName,
+      async job => {
+        expect(job.data.foo).to.be.equal('bar');
+        await delay(250);
+      },
+      { connection, prefix, concurrency: 100 },
+    );
+    await worker.waitUntilReady();
+
+    // Add spy to worker.moveToActive
+    const spy = sinon.spy(worker, 'moveToActive');
+    const bclientSpy = sinon.spy(
+      await worker.blockingConnection.client,
+      'bzpopmin',
+    );
+
+    for (let i = 0; i < numJobs; i++) {
+      const job = await queue.add('test', { foo: 'bar' });
+      expect(job.id).to.be.ok;
+      expect(job.data.foo).to.be.eql('bar');
+    }
+
+    expect(bclientSpy.callCount).to.be.equal(1);
+
+    await new Promise<void>((resolve, reject) => {
+      worker.on('completed', (job: Job, result: any) => {
+        completedJobs++;
+        if (completedJobs == numJobs) {
+          resolve();
+        }
+      });
+    });
+
+    // Check moveToActive was called numJobs + 2 times
+    expect(spy.callCount).to.be.equal(numJobs + 2);
+    expect(bclientSpy.callCount).to.be.equal(3);
+
+    await worker.close();
+  });
+
   it('does not process a job that is being processed when a new queue starts', async () => {
     this.timeout(12000);
     let err;
@@ -525,13 +614,17 @@ describe('workers', function () {
     expect(job.id).to.be.ok;
     expect(job.data.foo).to.be.eql('bar');
 
-    const failing = new Promise<void>(resolve => {
+    const failing = new Promise<void>((resolve, reject) => {
       worker.once('failed', async (job, err) => {
-        expect(job).to.be.ok;
-        expect(job.finishedOn).to.be.a('number');
-        expect(job.data.foo).to.be.eql('bar');
-        expect(err).to.be.eql(jobError);
-        resolve();
+        try {
+          expect(job).to.be.ok;
+          expect(job!.finishedOn).to.be.a('number');
+          expect(job!.data.foo).to.be.eql('bar');
+          expect(err).to.be.eql(jobError);
+          resolve();
+        } catch (err) {
+          reject(err);
+        }
       });
     });
 
@@ -752,12 +845,16 @@ describe('workers', function () {
       /* Try to gracefully close while having a job that will be completed running */
       worker.close();
 
-      await new Promise<void>(resolve => {
+      await new Promise<void>((resolve, reject) => {
         worker.once('completed', async job => {
-          expect(job).to.be.ok;
-          expect(job.finishedOn).to.be.a('number');
-          expect(job.data.foo).to.be.eql('bar');
-          resolve();
+          try {
+            expect(job).to.be.ok;
+            expect(job.finishedOn).to.be.a('number');
+            expect(job.data.foo).to.be.eql('bar');
+            resolve();
+          } catch (err) {
+            reject(err);
+          }
         });
       });
 
@@ -2120,7 +2217,7 @@ describe('workers', function () {
       const worker = new Worker(
         queueName,
         async job => {
-          if (job.attemptsMade !== 3) {
+          if (job.attemptsMade !== 2) {
             throw new Error('error');
           }
           return delay(100);
@@ -2292,7 +2389,7 @@ describe('workers', function () {
         const worker = new Worker(
           queueName,
           async job => {
-            expect(job.attemptsMade).to.equal(1);
+            expect(job.attemptsMade).to.equal(0);
             job.discard();
             throw new Error('unrecoverable error');
           },
@@ -2327,9 +2424,9 @@ describe('workers', function () {
       const worker = new Worker(
         queueName,
         async job => {
-          tries++;
           expect(job.attemptsMade).to.be.eql(tries);
-          if (job.attemptsMade < 2) {
+          tries++;
+          if (job.attemptsMade < 1) {
             throw new Error('Not yet!');
           }
         },
@@ -2363,10 +2460,10 @@ describe('workers', function () {
             async job => {
               id++;
               await delay(200);
-              if (job.attemptsMade === 1) {
+              if (job.attemptsMade === 0) {
                 expect(job.id).to.be.eql(`${id}`);
               }
-              if (job.id == '1' && job.attemptsMade < 2) {
+              if (job.id == '1' && job.attemptsMade < 1) {
                 throw new Error('Not yet!');
               }
             },
@@ -2411,11 +2508,11 @@ describe('workers', function () {
             queueName,
             async job => {
               await delay(200);
-              if (job.attemptsMade === 1) {
+              if (job.attemptsMade === 0) {
                 id++;
                 expect(job.id).to.be.eql(`${id}`);
               }
-              if (job.id == '1' && job.attemptsMade < 2) {
+              if (job.id == '1' && job.attemptsMade < 1) {
                 throw new Error('Not yet!');
               }
             },
@@ -2462,7 +2559,7 @@ describe('workers', function () {
         queueName,
         async job => {
           tries++;
-          if (job.attemptsMade < 4) {
+          if (job.attemptsMade < 3) {
             throw new Error('Not yet!');
           }
           expect(job.attemptsMade).to.be.eql(tries);
@@ -2504,7 +2601,7 @@ describe('workers', function () {
       const worker = new Worker(
         queueName,
         async job => {
-          if (job.attemptsMade < 3) {
+          if (job.attemptsMade < 2) {
             throw new Error('Not yet!');
           }
         },
@@ -2541,10 +2638,10 @@ describe('workers', function () {
         const worker = new Worker(
           queueName,
           async job => {
-            if (job.attemptsMade < 2) {
+            if (job.attemptsMade < 1) {
               throw new Error('Not yet!');
             }
-            if (job.attemptsMade < 3) {
+            if (job.attemptsMade < 2) {
               throw new UnrecoverableError('Unrecoverable');
             }
           },
@@ -2609,7 +2706,7 @@ describe('workers', function () {
                   break;
                 }
                 case Step.Second: {
-                  if (job.attemptsMade < 3) {
+                  if (job.attemptsMade < 2) {
                     throw new Error('Not yet!');
                   }
                   await job.updateData({
@@ -2781,7 +2878,8 @@ describe('workers', function () {
               const elapse = Date.now() - start;
               expect(elapse).to.be.greaterThan(200);
               expect(job.returnvalue).to.be.eql(Step.Finish);
-              expect(job.attemptsMade).to.be.eql(2);
+              expect(job.attemptsMade).to.be.eql(1);
+              expect(job.attemptsStarted).to.be.eql(2);
               resolve();
             });
 
@@ -2824,7 +2922,7 @@ describe('workers', function () {
                       { foo: 'bar' },
                       {
                         parent: {
-                          id: job.id,
+                          id: job.id!,
                           queue: job.queueQualifiedName,
                         },
                       },
@@ -2911,6 +3009,129 @@ describe('workers', function () {
           await childrenWorker.close();
           await parentQueue.close();
         });
+
+        describe('when skip attempt option is provided as true', () => {
+          it('should wait children as one step of the parent job whithout incrementing attemptMade', async function () {
+            this.timeout(8000);
+            const parentQueueName = `parent-queue-${v4()}`;
+            const parentQueue = new Queue(parentQueueName, {
+              connection,
+              prefix,
+            });
+
+            enum Step {
+              Initial,
+              Second,
+              Third,
+              Finish,
+            }
+
+            let waitingChildrenStepExecutions = 0;
+
+            const worker = new Worker(
+              parentQueueName,
+              async (job, token) => {
+                let step = job.data.step;
+                while (step !== Step.Finish) {
+                  switch (step) {
+                    case Step.Initial: {
+                      await queue.add(
+                        'child-1',
+                        { foo: 'bar' },
+                        {
+                          parent: {
+                            id: job.id!,
+                            queue: job.queueQualifiedName,
+                          },
+                        },
+                      );
+                      await job.updateData({
+                        step: Step.Second,
+                      });
+                      step = Step.Second;
+                      break;
+                    }
+                    case Step.Second: {
+                      await queue.add(
+                        'child-2',
+                        { foo: 'bar' },
+                        {
+                          parent: {
+                            id: job.id!,
+                            queue: `${prefix}:${parentQueueName}`,
+                          },
+                        },
+                      );
+                      await job.updateData({
+                        step: Step.Third,
+                      });
+                      step = Step.Third;
+                      break;
+                    }
+                    case Step.Third: {
+                      waitingChildrenStepExecutions++;
+                      const shouldWait = await job.moveToWaitingChildren(
+                        token!,
+                      );
+                      if (!shouldWait) {
+                        await job.updateData({
+                          step: Step.Finish,
+                        });
+                        step = Step.Finish;
+                        return Step.Finish;
+                      } else {
+                        throw new WaitingChildrenError();
+                      }
+                    }
+                    default: {
+                      throw new Error('invalid step');
+                    }
+                  }
+                }
+              },
+              { connection, prefix },
+            );
+            const childrenWorker = new Worker(
+              queueName,
+              async () => {
+                await delay(200);
+              },
+              {
+                connection,
+                prefix,
+              },
+            );
+            await childrenWorker.waitUntilReady();
+            await worker.waitUntilReady();
+
+            await parentQueue.add(
+              'test',
+              { step: Step.Initial },
+              {
+                attempts: 3,
+                backoff: 1000,
+              },
+            );
+
+            await new Promise<void>((resolve, reject) => {
+              worker.on('completed', job => {
+                expect(job.returnvalue).to.equal(Step.Finish);
+                expect(job.attemptsMade).to.be.eql(1);
+                expect(job.attemptsStarted).to.be.eql(2);
+                resolve();
+              });
+
+              worker.on('error', () => {
+                reject();
+              });
+            });
+
+            expect(waitingChildrenStepExecutions).to.be.equal(2);
+            await worker.close();
+            await childrenWorker.close();
+            await parentQueue.close();
+          });
+        });
       });
     });
 
@@ -2920,7 +3141,7 @@ describe('workers', function () {
       const worker = new Worker(
         queueName,
         async job => {
-          if (job.attemptsMade < 3) {
+          if (job.attemptsMade < 2) {
             throw new Error('Not yet!');
           }
         },
@@ -2960,7 +3181,7 @@ describe('workers', function () {
       const worker = new Worker(
         queueName,
         async job => {
-          if (job.attemptsMade < 3) {
+          if (job.attemptsMade < 2) {
             throw new Error('Not yet!');
           }
         },
@@ -3007,7 +3228,7 @@ describe('workers', function () {
         const worker = new Worker(
           queueName,
           async job => {
-            if (job.attemptsMade < 3) {
+            if (job.attemptsMade < 2) {
               throw new Error('Not yet!');
             }
           },
@@ -3089,7 +3310,7 @@ describe('workers', function () {
         queueName,
         async job => {
           tries++;
-          if (job.attemptsMade < 3) {
+          if (job.attemptsMade < 2) {
             throw new Error('Not yet!');
           }
         },
@@ -3137,7 +3358,7 @@ describe('workers', function () {
       const worker = new Worker(
         queueName,
         async job => {
-          if (job.attemptsMade < 3) {
+          if (job.attemptsMade < 2) {
             throw new CustomError('Hey, custom error!');
           }
         },
@@ -3250,7 +3471,7 @@ describe('workers', function () {
       const worker = new Worker(
         queueName,
         async (job: Job) => {
-          if (job.attemptsMade < 3) {
+          if (job.attemptsMade < 2) {
             throw new Error('some error');
           }
         },
@@ -3563,19 +3784,19 @@ describe('workers', function () {
 
         await Job.create(queue, 'testJob1', values[0], {
           parent: {
-            id: parent.id,
+            id: parent.id!,
             queue: `${prefix}:${parentQueueName}`,
           },
         });
         await Job.create(queue, 'testJob2', values[1], {
           parent: {
-            id: parent.id,
+            id: parent.id!,
             queue: `${prefix}:${parentQueueName}`,
           },
         });
         await Job.create(queue, 'testJob3', values[2], {
           parent: {
-            id: parent.id,
+            id: parent.id!,
             queue: `${prefix}:${parentQueueName}`,
           },
         });
@@ -3597,7 +3818,7 @@ describe('workers', function () {
           parentToken,
           {
             child: {
-              id: child3.id,
+              id: child3.id!,
               queue: `${prefix}:${queueName}`,
             },
           },